refactor(interactive): extract agent_log formatting, drop standalone visualizer
The standalone visualizer/visualizer.py is superseded by the web UI; only
its transcript-formatting (TRANSCRIPT_FILES, format_entry) was still used,
imported via a sys.path hack. Move that into a first-class
src/interactive/agent_log.py module and import it directly; delete the
unused standalone viewer (CLI + HTTP server) and update the web docs.
